+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## Changelog
 ### 2.0.0
 * Rebuilt plugin with new Settings API, everything is now 100% customizable (located in "Settings" > "Blogging")
-* Profile settings page relocated to "Settings" > "Blogging"
-* Reorder Pages menu further up (thanks to [Piet Bos](https://github.com/factmaven/disable-blogging/pull/2))
+* Profile settings page relocated to "Settings" > "Blogging" (thanks [cjbj](https://wordpress.org/support/profile/cjbj))
+* Reorder Pages menu further up (thanks to [Piet Bos](https://wordpress.org/support/users/senlin/))
 * Force changes to the following options in the settings:
 ** **Discussion**: Attempt to notify any blogs linked to from the article (unchecked)
 * Various code improvement and performance